[수정] title 컴포넌트 공통 스타일 분리하여 조건부 렌더링 최소화
