+++ frontend/src/components/FriendSearch/index.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import { ReactComponent as FriendShape } from '../../assets/shape/friend_shape.svg';
+
 const FriendSearchWrapper = styled.article`
   width: 100%;
   height: 77.3%;
@@ -19,9 +21,18 @@ const FriendSearchContainer = styled.div`
   }
 `;
 
+const FreindGreenShpate = styled(FriendShape)`
+  position: fixed;
+  width: 250px;
+  height: 400px;
+  top: 20px;
+  right: 40px;
+`;
+
 function FriendSearch({ children }) {
   return (
     <FriendSearchWrapper>
+      <FreindGreenShpate />
       <FriendSearchContainer>{children}</FriendSearchContainer>
     </FriendSearchWrapper>
   );
